Move server tools into subpackage.
