+++ src/mcp/server/tools/tools.ts
